requirements.txt
flask
flask-cors
pillow
numpy<2
requests
gunicorn
python-dotenv